requirements.txt
numpy
matplotlib
pandas
scikit-learn
umap-learn
wheel
seaborn